+++ src/Knet.jl
@@ -2,39 +2,23 @@ module Knet
 using Libdl
 # using LinearAlgebra, Statistics, SpecialFunctions, Libdl
 
-# To see debug output, set DBGFLAGS to non-zero. Each bit of DBGFLAGS
-# can be used to show a subset of dbg messages indicated by the `bit`
-# argument to the `dbg` macro.
-const DBGFLAGS = 0
-macro dbg(bit,x); if (1<<bit) & DBGFLAGS != 0; esc(:(println(_dbg($x)))); end; end;
-
-# To perform profiling, set PROFILING to true. (moved this to gpu.jl)
-# const PROFILING = false
-# macro gs(); if PROFILING; esc(:(ccall(("cudaDeviceSynchronize","libcudart"),UInt32,()))); end; end
+# To see debug output, start julia with `JULIA_DEBUG=Knet julia`
+# To perform profiling, set ENV["KNET_TIMER"] to "true" and rebuild Knet. (moved this to gpu.jl)
+# The @dbg macro below evaluates `ex` only when debugging. The @debug macro prints stuff as documented in Julia.
+macro dbg(ex); :(if Base.CoreLogging.current_logger_for_env(Base.CoreLogging.Debug,:none,Knet)!==nothing; $(esc(ex)); end); end
 
 const libknet8 = Libdl.find_library(["libknet8"], [joinpath(dirname(@__DIR__),"deps")])
 
-using AutoGrad
-using AutoGrad: forw, back, Tape
-using AutoGrad: Rec             # TODO; deprecate after AutoGrad 1.1
-export grad, value, Param
-export gradloss, getval         # TODO: deprecate after AutoGrad 1.1?
-if isdefined(AutoGrad,:Param); @eval begin
-    using AutoGrad: Value
-    export @diff
-end; else; @eval begin          # TODO: deprecate after AutoGrad 1.1
-    const value = getval
-    const Value = Rec
-    const Param = Rec
-    macro diff(x) :(throw(UndefVarError(:@diff))) end
-end; end
+using  AutoGrad: @diff, Param, params, grad, gradloss, value, cat1d, @primitive, @zerograd, @primitive1, @zerograd1, forw, back, Value, AutoGrad
+export AutoGrad, @diff, Param, params, grad, gradloss, value, cat1d #@primitive, @zerograd, @primitive1, @zerograd1, forw, back, Value, getval
 
 include("gpu.jl");              export gpu
 include("uva.jl")
 include("kptr.jl");             export knetgc # KnetPtr
 include("karray.jl");           export KnetArray
+include("gcnode.jl");
 include("ops.jl");
-include("unary.jl");            export relu, sigm, invx
+include("unary.jl");            export relu, sigm, invx, elu, selu
 include("broadcast.jl");        # elementwise broadcasting operations
 include("reduction.jl");        # sum, max, mean, etc.
 include("linalg.jl");           export mat # matmul, axpy!, transpose, (i)permutedims
@@ -44,13 +28,15 @@ include("batchnorm.jl");        export batchnorm, bnmoments, bnparams
 include("rnn.jl");              export rnnforw, rnninit, rnnparam, rnnparams, RNN # TODO: deprecate old interface
 include("data.jl");             export Data, minibatch
 include("model.jl");		export param, param0, train!, Train
-include("loss.jl");             export logp, logsumexp, nll, accuracy, zeroone # TODO: PR
+include("loss.jl");             export logp, logsoftmax, logsumexp, softmax, nll, logistic, bce, accuracy, zeroone # TODO: PR
 include("dropout.jl");          export dropout
 include("update.jl"); 		export SGD, Sgd, Momentum, Nesterov, Adam, Adagrad, Adadelta, Rmsprop, update!, optimizers
 include("distributions.jl"); 	export gaussian, xavier, bilinear
 include("random.jl");           export setseed  # TODO: deprecate setseed
 include("hyperopt.jl");         export hyperband, goldensection
-include("jld.jl");              export RnnJLD,KnetJLD
+include("serialize.jl");        export gpucopy,cpucopy
+include("jld.jl");              # load, save, @load, @save; not exported use with Knet. prefix.
+
 
 """
     Knet.dir(path...)